Add endpoint to api
