+++ requirements.txt
@@ -1,0 +1,29 @@
+uvicorn~=0.18.3
+fastapi~=0.85.0
+starlette~=0.20.4
+matplotlib>=3.2.2
+numpy>=1.18.5
+opencv-python>=4.1.1
+Pillow>=7.1.2
+PyYAML>=5.3.1
+requests>=2.23.0
+scipy>=1.4.1
+tqdm>=4.41.0
+protobuf<4.21.3
+tensorboard>=2.4.1
+pandas>=1.1.4
+seaborn>=0.11.0
+ipython  # interactive notebook
+psutil  # system utilization
+thop~=0.1.1-2209072238
+dependency_injector==4.40.0
+
+scikit-image~=0.19.2
+filterpy~=1.4.5
+moviepy
+firebase-admin
+
+ffmpeg_python>=0.2.0
+python_hostlist>=1.21
+timm>=0.4.12
+transformers>=4.31.0
