+++ requirements_gpu.txt
@@ -1,0 +1,7 @@
+# For Torch GPU
+
+-i  https://download.pytorch.org/whl/cu113
+
+torch==1.11.0+cu113
+torchvision==0.12.0+cu113
+
